update directory for pattern folder
