+++ springbootWithPostgresql/src/main/java/com/example/springbootWithPostgresql/pattern/template/UpdateUserTemplate/UpdateTheme.java
@@ -1,4 +1,4 @@
-package com.example.springbootWithPostgresql.pattern.UpdateUserTemplate;
+package com.example.springbootWithPostgresql.pattern.template.UpdateUserTemplate;
 
 import org.springframework.stereotype.Component;
 
@@ -6,15 +6,15 @@ import com.example.springbootWithPostgresql.entity.AccountEntity;
 import com.example.springbootWithPostgresql.repository.AccRepo;
 
 @Component
-public class UpdatePassword extends UpdateUserTemplate {
-    public UpdatePassword(AccRepo accRepo) {
+public class UpdateTheme extends UpdateUserTemplate {
+    public UpdateTheme(AccRepo accRepo) {
         super(accRepo);
     }
 
     @Override
     protected void updateAttributes(AccountEntity source, AccountEntity destination) {
-        if (!source.getPassword().isEmpty()) {
-            destination.setPassword(source.getPassword());
+        if (source.getHasdarktheme() != null) {
+            destination.setHasdarktheme(source.getHasdarktheme());
         }
     }
 }
